LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/cache-storage-reporting-shared-worker.https-expected.txt


FAIL COEP support on SharedWorker. assert_equals: expected 1 but got 0
FAIL COEP-Report-Only support on SharedWorker. assert_equals: expected 1 but got 0
